Working my way toward a consistent interface fo...
Working my way toward a consistent interface for scala.io.Source.
Now all the Source factory methods whose input is something Byte
based rather than Char based require an implicit Codec, which will be
Codec.default if no implicit is available. The idea is that if one uses
the scala APIs, your desired encoding can be expressed as an implicit
value and it will always be used; and that we (if so inclined) can
change the default in one place.