This is the beginning of a preventative effort ...
This is the beginning of a preventative effort to keep the collections
organization from drifting too far away from internal consistency. For
the various Proxy and Forwarder classes to be reliable and not missing
methods here and there, we need all methods defined in a canonical
location. I started organizing in collection/interfaces for the time
being.