Took a swing at sorting out sorting.
rewriting the Sorting methods to accept Orderings and adding a sorted
method to SeqLike, because we should all be pretty tired of writing
".sortWith(_ < _)" by now. I think it should be called "sort", not
"sorted", but that refuses to coexist gracefully with the deprecated
sort in List.

Review by moors (chosen pretty arbitrarily, someone at epfl should
review it but I don't know who deserves the nomination.)