Discovered and disproved this unlikely truth:
  scala> (1 to 1 drop 1) == (1 to 1)
  res0: Boolean = true

It was introduced in r21349 which was to fix scala/scala#2535, but led to scala/scala#3529.
I humbly suggest we can't afford to introduce bugs of this severity in
the pursuit of corner cases such as Ranges which use Int.MaxValue as a
boundary. And looking at the patch I find the "after" code a lot harder
to follow. Closes scala/scala#3529. Review by prokopec.