Overrode addString in TraversableViewLike.
the Transform-derived traits within view resisted evaluating the entire
sequence on a toString call, but the original view returned from a call
to .view did not. This has a particularly bad result in the case of
Stream, as for instance:

  Stream from 1 view

would enter infiniteloopiland in the repl despite the fact that it
should be doubly resistant to eager evaluation.

Review by prokopec.