Two improvements (or so I claim) to MatchErrors.
toString when constructing the exception: it need not be called
until/unless it is printed, which it may never be. 2) Include the name
of the class which triggered it. Don't tell me you haven't wanted that
as many times as I have.

(Sidebar on commit message semantics: I assume nobody interprets "no
review" to mean anything like "unreviewable decision!" It only means I'm
pretty sure the code does what I intended. The floor is always open.)

That said, no review.