Added separate bitmaps for private and transien...
Added separate bitmaps for private and transient lazy vals. Closes
scala/scala#3038, scala/scala#1573. Review by dragos. I had to fix a couple of initialization
issues that checkinit forced me to do and that weren't a problem before
because the bitmap was serialized even for @transitive. For that I
needed to change the setters in checkinit so that they also update the
bitmap.