Discovered bug in mapConserve introduced in r23...
Discovered bug in mapConserve introduced in r23038 when a tail recursive
version was submitted. I am optimistic this is a player in recent
performance degradation based on the timing of the commit and the fact
that I was examining mapConserve because of info from my homemade
profiler. (But it's late and I only just found it, so I'll be the last
to know.) Review by odersky.