Override checks and self-types still needed som...
Override checks and self-types still needed some work (ticket scala/scala#2808.)
I believe this is the right change. The modifications in parallel and
actors were a result of the files failing to compile after the change to
allOverriddenSymbols. (I am taking the position that that aspect is a
feature.) Review by malayeri, odersky.