A test case for recently fixed scala/scala#4114.  Plus!
I had closed scala/scala#2441 as a duplicate of that, but unfortunately scala/scala#4114 did
not bring scala/scala#2441 along with it. Then I realized I'm a programmer, not a
helpless trac watcher. As is often the case with thes things, fixing
that revealed a bug in the library. Closes scala/scala#2441 for real, review by
odersky.