Rediscovering that transpose sometimes throws a...
Rediscovering that transpose sometimes throws an exception on irregular
lists and sometimes returns an irregular result (depending on whether
the first collection is the longest or not) indicated that this needs a
better resolution than the status quo. Determination: as long as we're
throwing an exception on any invalid input, we should throw an exception
on all invalid input, so that's what it does now. I postpone any attempt
to offer a variation accepting a "hole value". Closes scala/scala#3597, review by
community.