As per discussion documented in SI-1799, brough...
As per discussion documented in SI-1799, brought back the ProductN
traits and synthesized them into case classes. It's -Xexperimental for
now because there may be minor implications for existing code which
should be discussed. And also because I snuck in another "improvement"
but it's probably too dangerous to be touching productIterator directly
and it should go into something else.

  scala> case class Bippy(x: Int, y: Int)
  defined class Bippy

  scala> Bippy(5, 10).productIterator
  res0: Iterator[Int] = non-empty iterator
                 ^^^----- as opposed to Iterator[Any]

There is an even better idea available than lubbing the case class field
types: it starts with "H" and ends with "List"...

Review by oderksy.