Working on jar creation infrastructure.
output generation (but only then, since otherwise we're not creating the
jar):

  1) -Xmain-class foo.Bar will give the jar a Main-Class of foo.Bar 2)
  Alternatively, if there is only one runnable program, that will be
  the Main-Class 3) Always, the jar's manifest will have an entry for
  Scala-Compiler-Version.

Not very relatedly, a warning is now issued when a module has a main
method but a runnable program will not be generated. Closes SI-4861.

This represents an opening step toward automatically recognizing
mismatched bytecode situations: coarse, but useful and safe. Review by
mirco.