Moved the classes and objects which are defined...
Moved the classes and objects which are defined in package objects out.

In principle this is something you should be able to do. In practice
right now it means bugs, to no advantage.

I also deprecated RangeUtils, an unused, undocumented trait in the
immutable package. It seems like there is a ton of stuff in the public
API which should not be in the public API. It's really tedious having to
go through a whole deprecation cycle to dispose of what could have been
an internal-only class (and was presumably intended as such given the
the absence of documentation.)

No review.