Reversed the values of "is" and "is not" in rec...
Reversed the values of "is" and "is not" in recent for comprehension
deprecation.

  DO NOT BLOW HATCH
  REPEAT
  DO NOT BLOW HATCH
  "Roger! Hatch blown."

Events reveal it was all baby, no bathwater. It turns out that the
specification is merely a document, not infallible holy writ as we
had all previously believed. So it is not the ABSENCE of val in a for
comprehension assignment which is deprecated, it is the PRESENCE of val.
Summarizing again, more accurately perhaps:

  for (x <- 1 to 5 ; y = x) yield x+y           // THAT's the one
  for (val x <- 1 to 5 ; y = x) yield x+y       // fail
  for (val x <- 1 to 5 ; val y = x) yield x+y   // fail
  for (x <- 1 to 5 ; val y = x) yield x+y       // deprecated

No review.