A conceivably pretty bad performance bug in bui...
A conceivably pretty bad performance bug in builders.

SI-4821 pointed out that ArrayBuffer's ++ checks for a cheap size method
by matching on IndexedSeq, but mutable.IndexedSeq, so all immutable
collections are thrown in the same group as linear seqs. I went looking
for other examples of this and found them, in key classes like Builder.

The "type shadowing trap" is a serious issue in the collections. Closes
SI-4821, no review.