Set fields in immutable hash maps and hash sets to vals.
This is part of an effort to make the immutable collections
(more) thread safe. The `::` still has non-final member fields
for head and tail, but there is not much that can be done right
now about that, since these fields are used by list buffers.

Tried writing a test with unsafe initialization, but could not
invent a scenario which actually fails, at least on the JDK6.