Introduce getAnnotations that triggers symbol completion
Default getter for annotations doesn't perform initialization, hence
we've faced the following bug: https://issues.scala-lang.org/browse/SI-5423.

One of the approaches to fixing it would be to auto-complete on getter,
but according to Martin we'd better not do that because of cycles.

That's why I'm just introducing a new, eager, variation of `annotations'
and redirecting public API to it.

Review by @odersky.