SIP-11 String Interpolation Simplification
Changed StringContext.f which implements the formatted
string interpolator.

Any '%' not in formatting position is left in the resulting string literally.
However, instead of adding %s format holders and extending the args with "%"
expressions, these '%' are replaced by '%%'. The formatter then converts '%%'
(percent formatter) to the literal '%' (\u0025). This also simplifies the spec.

The interpolation tests still pass.