Fixes SI-5201.  Adds flatten to TraversableViewLike.
Note: This commit exposes a pretty rich type on flatten in views.  HOWEVER, because we don't
capture the higher kinded type of the underlying collection, it makes returning a more minimal type
pretty dang hard.  I can imagine a very breaking and painful change of capturing the underling
collection as a higher-kinded type as well as the current view type in a *ViewLike.scala.

I hope this kind of issue, along with others, drives a rethink of our view API design.