Avoid relying on auto-tupling.
When you call

  Some(a, b, c)

The compiler does eventually figure out that you meant

  Some((a, b, c))

However the compiler finds this inordinately taxing, and
because this auto-tupling is dangerous, -Xlint warns about it.
In the interests of being able to hear -Xlint's important
points, we write Some((a, b, c)) to give it less to warn about.